working Profile fixturator and fixturator.rs moved to root src folder
