+++ dna/tests/src/lib.rs
@@ -1,3 +1,4 @@
 mod project;
 mod profiles;
 mod test_lib;
+mod fixtures;
