+++ dna/tests/src/project/goal_vote.rs
@@ -1,6 +1,6 @@
 #[cfg(test)]
 pub mod tests {
-    use crate::project::fixtures::fixtures::{
+    use crate::fixtures::fixtures::{
         GoalVoteFixturator, WrappedAgentPubKeyFixturator,
         WrappedHeaderHashFixturator,
     };
