+++ dna/tests/src/project/mod.rs
@@ -6,5 +6,4 @@ pub mod goal_member;
 pub mod goal_vote;
 pub mod member;
 pub mod project_meta;
-pub mod fixtures;
 pub mod project_init;
